Move testing directory and add information to pyproject.toml
